fix: リントエラーを解消し、モジュールを分割
