+++ app/core/pyproject.toml
@@ -6,6 +6,7 @@ readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
     "open3d>=0.19.0",
+    "psutil>=7.2.2",
     "trimesh>=4.11.1",
 ]
