+++ app/core/src/core/utils/profiler/__init__.py
@@ -1,0 +1,7 @@
+"""Performance profiling utilities for RANSAC optimization.
+
+このモジュールは、RANSACアルゴリズムの性能を測定・分析するためのプロファイリングツールを提供します。
+関数レベルのタイミング測定、メモリ使用量の追跡、統計レポートの生成が可能です。
+"""
+
+__all__ = ["profile", "profile_block"]
